+++ geolib/models/dfoundations/__init__.py
@@ -1,1 +1,5 @@
+"""
+The D-Series D-Foundations model definition.
+"""
+
 from .dfoundations_model import DFoundationsModel