+++ geolib/models/dsheetpiling/settings.py
@@ -33,12 +33,12 @@ class LateralEarthPressureMethodStage(IntEnum):
 class ModelType(IntEnum):
     """Represents the model types from D-Sheet Piling using the same integer values used in the application.
 
-    The Model type determines the kind of calculation that's performed on the construction.
+    The Model type determines the kind of construction: single pile or sheet piling.
+    Note that the model type SHEET_PILING includes sheet piling, wooden sheet piling and diaphragm wall.
     """
 
     SHEET_PILING = 0
     SINGLE_PILE = 1
-    DIAPHRAGM_WALL = 2
 
 
 class SinglePileLoadOptions(Enum):