+++ geolib/service/__init__.py
@@ -1,0 +1,3 @@
+"""
+Module containing web service related code.
+"""